moved api-model folder to msp-core - builds & serves
