+++ src/app/modules/benefit/services/msp-api-benefit.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { MspLogService } from '../../../services/log.service';
 import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
 import { BenefitApplication } from '../models/benefit-application.model';
-import { AttachmentType, _ApplicationTypeNameSpace } from '../../enrolment/pages/api-model/applicationTypes';
+import { AttachmentType, _ApplicationTypeNameSpace } from '../../msp-core/api-model/applicationTypes';
 import { environment } from '../../../../environments/environment';
 import * as moment from 'moment';
 import { AbstractHttpService } from 'moh-common-lib';
