+++ src/app/modules/msp-core/api-model/applicationTypes.ts
