+++ src/app/modules/msp-core/api-model/xml-primitives.d.ts
