+++ src/app/modules/msp-core/api-model/xsd/ApplicationTypes.xsd
