+++ src/app/modules/msp-core/api-model/xsd/AssistanceTypes.xsd
