+++ src/app/modules/msp-core/api-model/xsd/NameAlignmentTypes.xsd
