+++ src/app/services/msp-maintenance.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
-import { _ApplicationTypeNameSpace } from '../modules/enrolment/pages/api-model/applicationTypes';
+import { _ApplicationTypeNameSpace } from '../modules/msp-core/api-model/applicationTypes';
 import { ISpaEnvResponse } from '../components/msp/model/spa-env-response.interface';
 import { MspLog2Service } from './log2.service';
 import * as moment from 'moment';
