+++ src/version.GENERATED.ts
@@ -1,9 +1,9 @@
 // DO NOT DELETE OR APP WILL FAIL TO COMPILE! Generated from version.js
-export const gitCommit = '17e16696';
-export const buildTime = '7/16/2019 at 10:50:49 AM';
+export const gitCommit = '4d5c98eb';
+export const buildTime = '7/16/2019 at 11:45:12 AM';
 /** App version retrieved from package.json. */
 export const projectVersion = '1.1.0';
 /** Human readable message  */
-export const message = 'Canonical Version: 17e16696 - 7/16/2019 at 10:50:49 AM (v1.1.0) \nhttps://github.com/bcgov/MyGovBC-MSP/commits/17e16696';
+export const message = 'Canonical Version: 4d5c98eb - 7/16/2019 at 11:45:12 AM (v1.1.0) \nhttps://github.com/bcgov/MyGovBC-MSP/commits/4d5c98eb';
 /** If true, other values should be present. */
 export const success = true;
